+++ CSharpGL/NewFontResource/TextureBuilder.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+
+namespace CSharpGL.NewFontResource
+{
+    public class TextureBuilder
+    {
+
+        public uint BuildTexture(Bitmap bitmap)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
